+++ GrosBrasInc/GrosBrasInc/ViewModels/ShoppingCartRemoveViewModel.cs
@@ -8,7 +8,7 @@ namespace GrosBrasInc.ViewModels
     public class ShoppingCartRemoveViewModel
     {
         public string Message { get; set; }
-        public float CartTotal { get; set; }
+        public double CartTotal { get; set; }
         public int CartCount { get; set; }
         public int ItemCount { get; set; }
         public int DeleteId { get; set; }
